Resolve pylint for whole blokus_rl module
