+++ blokus_rl/blokus/game/__init__.py
@@ -1,0 +1,2 @@
+from .blokus_game import BlokusGame
+from .board import Board
