+++ blokus_rl/blokus/model/blokus_nnet.py
@@ -1,8 +1,4 @@
 """Deep neural network model for PPO algorithm."""
-import os
-import time
-from pathlib import Path
-
 import numpy as np
 import torch
 import torch.nn as nn
