+++ blokus_rl/blokus/players/__init__.py
@@ -1,0 +1,1 @@
+from .player import AiPlayer, Player
