+++ blokus_rl/blokus/players/user_player.py
@@ -1,3 +1,4 @@
+# pylint: skip-file
 import matplotlib.pyplot as plt
 
 from ...hparams import MCTSHparams
