+++ blokus_rl/blokus/shapes/__init__.py
@@ -1,0 +1,1 @@
+from .shapes import get_all_shapes, Shape
