--- blokus_rl/stablebaseline/__init__.py
@@ -1,1 +1,0 @@
-from .sb_ppo import SBTrainer
